修复ub-pkg-mem包的依赖关系、增加新的驱动列表及docker容器中的检查

+++ ub-pkg-manager.spec
@@ -3,12 +3,12 @@ Version:        0.0.4
 Release:        6
 Summary:        The full function of UB OS Component
 
-License:        mulan
+License:        MulanPSL2
 Source0:        %{name}-%{version}.tar.gz
 
 ExclusiveArch:  aarch64
 BuildRequires:  python3-setuptools
-Requires:       python3 >= 3.7 python3-rich python3-pyyaml
+Recommends:       python3 >= 3.6 python3-rich python3-pyyaml
 Requires:       systemd
 Requires:       ub-pkg-urma = %{version}-%{release}
 Requires:       ub-pkg-mem = %{version}-%{release}
@@ -31,7 +31,7 @@ user-space software packages.
 # Subpackage: ub-pkg-mem
 %package -n ub-pkg-mem
 Summary:        The UnifiedBus Pooled Memory function software package management tool of UB OS Component
-Requires:       ub-pkg-urma = %{version}-%{release} sysSentry libobmm
+Requires:       sysSentry libobmm ubctl ubutils libummu libcdma
 
 %description -n ub-pkg-mem
 UnifiedBus Pooled Memory function software package management tool, which 
@@ -41,7 +41,7 @@ user-space software packages.
 # Subpackage: ub-pkg-virt
 %package -n ub-pkg-virt
 Summary:        The UnifiedBus Device Virtualization function software package management tool of UB OS Component
-Requires:       ub-pkg-urma = %{version}-%{release} qemu libvirt memlinkd
+Requires:       ub-pkg-urma = %{version}-%{release} qemu libvirt memlinkd edk2-aarch64
 
 %description -n ub-pkg-virt
 UnifiedBus Device Virtualization function software package management tool, 
@@ -49,6 +49,13 @@ which includes loading the kernel drivers required for device virtualization
 and installing user-space software packages.
 
 %global debug_package %{nil}
+%global is_not_container() %(cat << 'EOF'
+is_not_container() {
+    ! grep -qE "docker|lxc|podman|kubepods" /proc/1/cgroup 2>/dev/null
+}
+EOF
+)
+
 %prep
 %setup -q
 
@@ -94,83 +101,92 @@ cp -r src/ub_manage/scripts/testkit/* %{buildroot}/usr/local/ub-pkg-manager/bin/
 chmod -R 755 %{buildroot}/usr/local/ub-pkg-manager/bin/testkit/
 
 %post -n ub-pkg-manager
+%is_not_container
 
-# Reload systemd
-/bin/systemctl daemon-reload
-
-# Enable service by default
-/bin/systemctl enable ub-pkg-manager.service 2>/dev/null || :
-
-# Start main service
-/bin/systemctl start ub-pkg-manager.service 2>/dev/null || :
+# Reload, enable and start service if not in container
+if is_not_container; then
+    /bin/systemctl daemon-reload
+    /bin/systemctl enable ub-pkg-manager.service 2>/dev/null || :
+    /bin/systemctl start ub-pkg-manager.service 2>/dev/null || :
+fi
 
 %preun -n ub-pkg-manager
-if [ $1 -eq 0 ]; then
-    # Final removal, stop and disable service
+%is_not_container
+
+if [ $1 -eq 0 ] && is_not_container; then
+    # Final removal, stop and disable service if not in container
     /bin/systemctl stop ub-pkg-manager.service 2>/dev/null || :
     /bin/systemctl disable ub-pkg-manager.service 2>/dev/null || :
 fi
 
 # Subpackage: ub-pkg-urma post install
 %post -n ub-pkg-urma
-# Reload systemd
-/bin/systemctl daemon-reload
+%is_not_container
 
-# Enable service by default
-/bin/systemctl enable ub-pkg-urma.service 2>/dev/null || :
-
-# Start service
-/bin/systemctl start ub-pkg-urma.service 2>/dev/null || :
+# Reload, enable and start service if not in container
+if is_not_container; then
+    /bin/systemctl daemon-reload
+    /bin/systemctl enable ub-pkg-urma.service 2>/dev/null || :
+    /bin/systemctl start ub-pkg-urma.service 2>/dev/null || :
+fi
 
 # Subpackage: ub-pkg-urma pre uninstall
 %preun -n ub-pkg-urma
-if [ $1 -eq 0 ]; then
-    # Final removal, stop and disable service
+%is_not_container
+
+if [ $1 -eq 0 ] && is_not_container; then
+    # Final removal, stop and disable service if not in container
     /bin/systemctl stop ub-pkg-urma.service 2>/dev/null || :
     /bin/systemctl disable ub-pkg-urma.service 2>/dev/null || :
 fi
 
 # Subpackage: ub-pkg-mem post install
 %post -n ub-pkg-mem
-# Reload systemd
-/bin/systemctl daemon-reload
+%is_not_container
 
-# Enable service by default
-/bin/systemctl enable ub-pkg-mem.service 2>/dev/null || :
-
-# Start service after its dependency
-/bin/systemctl start ub-pkg-mem.service 2>/dev/null || :
+# Reload, enable and start service if not in container
+if is_not_container; then
+    /bin/systemctl daemon-reload
+    /bin/systemctl enable ub-pkg-mem.service 2>/dev/null || :
+    /bin/systemctl start ub-pkg-mem.service 2>/dev/null || :
+fi
 
 # Subpackage: ub-pkg-mem pre uninstall
 %preun -n ub-pkg-mem
-if [ $1 -eq 0 ]; then
-    # Final removal, stop and disable service
+%is_not_container
+
+if [ $1 -eq 0 ] && is_not_container; then
+    # Final removal, stop and disable service if not in container
     /bin/systemctl stop ub-pkg-mem.service 2>/dev/null || :
     /bin/systemctl disable ub-pkg-mem.service 2>/dev/null || :
 fi
 
 # Subpackage: ub-pkg-virt post install
 %post -n ub-pkg-virt
-# Reload systemd
-/bin/systemctl daemon-reload
+%is_not_container
 
-# Enable service by default
-/bin/systemctl enable ub-pkg-virt.service 2>/dev/null || :
-
-# Start service after its dependency
-/bin/systemctl start ub-pkg-virt.service 2>/dev/null || :
+# Reload, enable and start service if not in container
+if is_not_container; then
+    /bin/systemctl daemon-reload
+    /bin/systemctl enable ub-pkg-virt.service 2>/dev/null || :
+    /bin/systemctl start ub-pkg-virt.service 2>/dev/null || :
+fi
 
 # Subpackage: ub-pkg-virt pre uninstall
 %preun -n ub-pkg-virt
-if [ $1 -eq 0 ]; then
-    # Final removal, stop and disable service
+%is_not_container
+
+if [ $1 -eq 0 ] && is_not_container; then
+    # Final removal, stop and disable service if not in container
     /bin/systemctl stop ub-pkg-virt.service 2>/dev/null || :
     /bin/systemctl disable ub-pkg-virt.service 2>/dev/null || :
 fi
 
 %postun -n ub-pkg-manager
-# Reload systemd if upgrading or removing
-if [ $1 -ge 1 ]; then
+%is_not_container
+
+# Reload systemd if upgrading or removing and not in container
+if [ $1 -ge 1 ] && is_not_container; then
     /bin/systemctl daemon-reload
 fi
 
@@ -193,22 +209,28 @@ fi
 
 # Subpackage: ub-pkg-urma post uninstall
 %postun -n ub-pkg-urma
-# Reload systemd if upgrading or removing
-if [ $1 -ge 1 ]; then
+%is_not_container
+
+# Reload systemd if upgrading or removing and not in container
+if [ $1 -ge 1 ] && is_not_container; then
     /bin/systemctl daemon-reload
 fi
 
 # Subpackage: ub-pkg-mem post uninstall
 %postun -n ub-pkg-mem
-# Reload systemd if upgrading or removing
-if [ $1 -ge 1 ]; then
+%is_not_container
+
+# Reload systemd if upgrading or removing and not in container
+if [ $1 -ge 1 ] && is_not_container; then
     /bin/systemctl daemon-reload
 fi
 
 # Subpackage: ub-pkg-virt post uninstall
 %postun -n ub-pkg-virt
-# Reload systemd if upgrading or removing
-if [ $1 -ge 1 ]; then
+%is_not_container
+
+# Reload systemd if upgrading or removing and not in container
+if [ $1 -ge 1 ] && is_not_container; then
     /bin/systemctl daemon-reload
 fi
 
@@ -245,6 +267,14 @@ fi
 /usr/local/ub-pkg-manager/bin/03-ub-pkg-virt.sh
 
 %changelog
+* Tue Apr 21 2026 gongzhengtang <gong_zhengtang@163.com> - 0.0.4-6
+- Add ipourma ko module
+* Mon Mar 23 2026 gongzhengtang <gong_zhengtang@163.com> - 0.0.4-5
+- Optimize the prompt messages for both EFI and BIOS situations
+* Fri Mar 21 2026 gongzhengtang <gong_zhengtang@163.com> - 0.0.4-4
+- Improve command prompts and fix file-saving error
+* Tues Mar 10 2026 gongzhengtang <gong_zhengtang@163.com> - 0.0.4-3
+- This update addresses and resolves a number of known issues that were present in version 0.0.4.
 * Tues Mar 3 2026 gongzhengtang <gong_zhengtang@163.com> - 0.0.4-2
 - Add load, dump,rollback and list commands
 * Thur Feb 5 2026 gongzhengtang <gong_zhengtang@163.com> - 0.0.4-1
